@@ -1,22 +1,22 @@
-const React = require('react')
-
 const placeholderRegex = /(\{[\d|\w]+\})/
 
 const formatString = (str, ...valuesForPlaceholders) => {
   let hasObject = false
   const res = str
     .split(placeholderRegex)
-    .filter(textPart => !!textPart)
+    .filter((textPart) => !!textPart)
     .map((textPart, index) => {
       if (textPart.match(placeholderRegex)) {
         const valueForPlaceholder = valuesForPlaceholders[textPart.slice(1, -1)]
 
         if (React.isValidElement(valueForPlaceholder)) {
           hasObject = true
-          return React.Children.toArray(valueForPlaceholder).map(component => ({
-            ...component,
-            key: index.toString()
-          }))
+          return React.Children.toArray(valueForPlaceholder).map(
+            (component) => ({
+              ...component,
+              key: index.toString(),
+            }),
+          )
         }
 
         return valueForPlaceholder
@@ -28,4 +28,4 @@ const formatString = (str, ...valuesForPlaceholders) => {
   return res.join('')
 }
 
-module.exports = formatString
+export default formatString
