@@ -1,10 +1,10 @@
-const createLanguages = require('../createLanguages')
-const formatString = require('../formatString')
+import createLanguages from '../createLanguages'
+import formatString from '../formatString'
 
 describe('Languages should be create registered and then updated correctly', () => {
   test('init two languages', () => {
     const en = {
-      hello: 'Hello, {0} {1}!!!'
+      hello: 'Hello, {0} {1}!!!',
     }
 
     const firstName = 'Huy'
@@ -15,7 +15,7 @@ describe('Languages should be create registered and then updated correctly', () 
     Languages.data.en = en
 
     expect(
-      Languages.formatString(Languages.hello, firstName, lastName)
+      Languages.formatString(Languages.hello, firstName, lastName),
     ).toEqual('Hello, Huy Tran!!!')
   })
 })
